+++ github/CodeScanRule.pyi
